Refactor and finish QR code scanner
