+++ package.json
@@ -12,6 +12,8 @@
   "dependencies": {
     "@ngraveio/bc-ur": "^1.1.13",
     "@react-native/new-app-screen": "0.83.1",
+    "@react-navigation/native": "^7.1.28",
+    "@react-navigation/native-stack": "^7.13.0",
     "keycard-sdk": "^3.1.3",
     "react": "19.2.0",
     "react-native": "0.83.1",
@@ -20,6 +22,7 @@
     "react-native-keycard": "github:choppu/react-native-keycard",
     "react-native-paper": "^5.15.0",
     "react-native-safe-area-context": "^5.5.2",
+    "react-native-screens": "^4.23.0",
     "react-native-vector-icons": "^10.3.0"
   },
   "devDependencies": {
