+++ src/theme.ts
@@ -1,0 +1,17 @@
+import {MD3DarkTheme} from 'react-native-paper';
+
+const theme = {
+  ...MD3DarkTheme,
+  colors: {
+    ...MD3DarkTheme.colors,
+    primary: '#bb86fc',
+    secondary: '#03dac6',
+    surface: '#1e1e1e',
+    surfaceVariant: '#2d2d2d',
+    background: '#121212',
+    onSurface: '#ffffff',
+    onSurfaceVariant: 'rgba(255,255,255,0.7)',
+  },
+};
+
+export default theme;
